Add message for duplicate places per event
So some Nationals results do have duplicate places (ties) because it is
allowed by the rules for places > 6, but this complicates things with
Invitationals that sometimes give out awards to places up to 8.

For consistency, SciolyFF will require that all ties be broken, but
programs designed to parse SciolyFF may want to keep this in mind and
even allow input with these duplicate places.

Will want to document the above somewhere besides a commit message.