Fully deprecate (i.e. remove) SciolyFF::Helper
Forgot to do this before releasing 0.7.0, but it'll be done in 0.7.1,
the release that will surely need to happen as bugs are found in 0.7.0
when using it on unosmium.org. Should really write a test suite one of
these days instead of relying on the website. ¯\_(ツ)_/¯

Edit: ignore the above, forgot that you actually need to push the gem to
release it, so deprecation will actually happen in 0.7.0